+++ src/Curiosity.Notifications.EMail/EmailNotificationChannel.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Curiosity.EMail;
@@ -6,8 +8,6 @@ using Microsoft.Extensions.Logging;
 
 namespace Curiosity.Notifications.EMail
 {
-    //todo Add notification post processor
-
     /// <summary>
     /// Channel for sending EMail notifications.
     /// Uses registered at IoC implementation of <see cref="IEMailSender"/> to send notifications.
@@ -16,9 +16,17 @@ namespace Curiosity.Notifications.EMail
     {
         private readonly IEMailSender _sender;
 
-        public EmailNotificationChannel(ILogger<EmailNotificationChannel> logger, IEMailSender sender) : base(logger, EmailNotification.Type)
+        private readonly IReadOnlyList<IEMailNotificationPostProcessor> _postProcessors;
+
+        public EmailNotificationChannel(
+            ILogger<EmailNotificationChannel> logger,
+            IEMailSender sender,
+            IEnumerable<IEMailNotificationPostProcessor> postProcessors) : base(logger, EmailNotification.Type)
         {
             _sender = sender ?? throw new ArgumentNullException(nameof(sender));
+
+            if (postProcessors == null) throw new ArgumentNullException(nameof(postProcessors));
+            _postProcessors = postProcessors.ToArray();
         }
 
         /// <inheritdoc />
@@ -26,6 +34,7 @@ namespace Curiosity.Notifications.EMail
         {
             if (notification == null) throw new ArgumentNullException(nameof(notification));
 
+            // send EMail
             var result = notification.ExtraParams == null
                 ? await _sender.SendAsync(
                     notification.Email,
@@ -40,6 +49,23 @@ namespace Curiosity.Notifications.EMail
                     notification.IsBodyHtml,
                     notification.ExtraParams,
                     cancellationToken);
+
+            // execute post processing
+            for (var i = 0; i < _postProcessors.Count; i++)
+            {
+                var postProcessor = _postProcessors[i];
+
+                try
+                {
+                    await postProcessor.ProcessAsync(notification, result, cancellationToken);
+                }
+                catch (Exception e)
+                {
+                    Logger.LogError(e, $"Error while post processing EMail notification. Reason: {e.Message}");
+                }
+            }
+
+            // throw exception on failed result
             if (!result)
                 throw new InvalidOperationException($"Sending email to \"{notification.Email}\" failed");
         }