+++ src/Curiosity.Notifications.SMS/SmsNotificationChannel.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Curiosity.SMS;
@@ -10,18 +12,44 @@ namespace Curiosity.Notifications.SMS
     {
         private readonly ISmsSender _smsSender;
 
-        public SmsNotificationChannel(ILogger logger, ISmsSender smsSender) : base(logger, SmsNotification.Type)
+        private readonly IReadOnlyList<ISmsNotificationPostProcessor> _postProcessors;
+
+        public SmsNotificationChannel(
+            ILogger<SmsNotificationChannel> logger,
+            ISmsSender smsSender,
+            IEnumerable<ISmsNotificationPostProcessor> postProcessors) : base(logger, SmsNotification.Type)
         {
+            if (postProcessors == null) throw new ArgumentNullException(nameof(postProcessors));
             _smsSender = smsSender ?? throw new ArgumentNullException(nameof(smsSender));
+
+            _postProcessors = postProcessors.ToArray();
         }
 
         protected override async Task ProcessNotificationAsync(SmsNotification notification, CancellationToken cancellationToken = default)
         {
             if (notification == null) throw new ArgumentNullException(nameof(notification));
 
+            // send SMS
             var result = notification.ExtraParams == null
                 ? await _smsSender.SendSmsAsync(notification.PhoneNumber, notification.Message, cancellationToken)
                 : await _smsSender.SendSmsAsync(notification.PhoneNumber, notification.Message, notification.ExtraParams, cancellationToken);
+
+            // execute post processing
+            for (var i = 0; i < _postProcessors.Count; i++)
+            {
+                var postProcessor = _postProcessors[i];
+
+                try
+                {
+                    await postProcessor.ProcessAsync(notification, result, cancellationToken);
+                }
+                catch (Exception e)
+                {
+                    Logger.LogError(e, $"Error while post processing SMS notification. Reason: {e.Message}");
+                }
+            }
+
+            // throw exception if something failed
             if (!result.IsSuccess)
                 throw new InvalidOperationException($"Sending SMS to \"{notification.PhoneNumber}\" failed.");
         }