+++ src/Curiosity.SMS.Smsc/ISmscSender.cs
@@ -1,7 +1,7 @@
 namespace Curiosity.SMS.Smsc
 {
     /// <summary>
-    /// Service for sendind SMS via smsc.ru.
+    /// Service for sending SMS via smsc.ru.
     /// </summary>
     public interface ISmscSender : ISmsSender
     {