+++ src/Curiosity.SMS.Smsc/SmscOptions.cs
@@ -11,25 +11,23 @@ namespace Curiosity.SMS.Smsc
     {
         private const string CanNotBeEmptyErrorDescription = "can not be empty";
 
-        private const string ZeroOrNegativeErrorDescription = "should be greater than 0";
-
         /// <summary>
         /// Login for Smsc.
         /// </summary>
-        public string SmscLogin { get; set; }
+        public string SmscLogin { get; set; } = null!;
 
         /// <summary>
         /// Pass for Smsc.
         /// </summary>
-        public string SmscPassword { get; set; }
+        public string SmscPassword { get; set; } = null!;
 
         /// <summary>
         /// Sender's name.
         /// </summary>
-        public string SmscSender { get; set; }
+        public string SmscSender { get; set; } = null!;
 
         /// <inheritdoc />
-        public IReadOnlyCollection<ConfigurationValidationError> Validate(string prefix = null)
+        public IReadOnlyCollection<ConfigurationValidationError> Validate(string? prefix = null)
         {
             var errors = new ConfigurationValidationErrorCollection(prefix);
             errors.AddErrorIf(String.IsNullOrEmpty(SmscLogin), nameof(SmscLogin), CanNotBeEmptyErrorDescription);